refactor(trusty-agents): split persona.rs test module into persona_tests.rs
persona.rs's SLOC (which the 500-line cap counts by filename, not by
#[cfg(test)]) crossed the cap once the #3208 scope-enforcement tests were
added. Extracts the inline mod tests { ... } into a sibling
persona_tests.rs via #[path = "persona_tests.rs"] mod tests, matching the
builder.rs/builder_tests.rs and tools/mod.rs/registry_tests.rs precedent
already used in this workspace. Behavior and test coverage are unchanged —
verified all 10 persona tests still pass post-split, and that
filter_persona_tool_names / persona_allowed_tools remain private (not pub)
since persona_tests.rs is a child module of persona and can see private
items of its parent without any visibility widening.
